Make type conversion explicit.
git-svn-id: https://sourceforge.llnl.gov/svn/repos/samrai_project/trunk/MASTER@1134 f45145c3-882e-4954-aa44-00140844b9ef

+++ source/SAMRAI/mesh/TreeLoadBalancer.C
@@ -405,7 +405,7 @@ TreeLoadBalancer::loadBalanceBoxLevel(
     */
    const double fanout_size = max_local_load/d_global_avg_load;
    const int number_of_cycles =
-      int(ceil( log(fanout_size)/log(d_max_cycle_spread_ratio) ));
+      int(ceil( log(fanout_size)/log(double(d_max_cycle_spread_ratio)) ));
       if (d_print_steps) {
          tbox::plog << "TreeLoadBalancer::loadBalanceBoxLevel"
                     << " max_cycle_spread_ratio=" << d_max_cycle_spread_ratio
